requirements.txt
Flask
requests
Pillow
timeloop
